Platform list uses the normalized values.

+++ src/view/widget_text.h
@@ -27,6 +27,6 @@ typedef struct {
 	gboolean shadow;
 } WidgetText;
 
-WidgetText* meh_widget_text_new(const Font* font, const char* text, int x, int y, SDL_Color color, gboolean shadow);
+WidgetText* meh_widget_text_new(const Font* font, const char* text, float x, float y, SDL_Color color, gboolean shadow);
 void meh_widget_text_destroy(WidgetText* text);
 void meh_widget_text_render(Window* window, const WidgetText* text);
